requirements.txt
beautifulsoup4
odfpy

